backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-genai==0.3.0
pybullet==3.2.6
pydantic==2.5.3
pillow==10.2.0
python-dotenv==1.0.1